Beginner
This code could be streamlined, but it is functioning.  After attempting to utilize previous code that was inconsistently structured, I rewrote this code in a few hours to help organize the flow and make more sense.  It is not perfect, and yes it can stand to be made more efficient. This bottom line is that this code is above the requirements and works.